addDataFrameToHTML {myHTMLtool}	R Documentation

Add data frame to HTML object.

Description

Convert a table data frame to a javascript object and create a .html file capable to loading the .json file

Usage

addDataFrameToHTML(dataframe, htmlObj, customFilter = "auto",
  fieldnames = NULL, tableParentNode = NULL)

Arguments

dataframe
The data frame containing the table to be added to the output .html file

htmlObj
A html object returned by initHTML()

customFilter
A user-defined criteria to add individual column filtering in the HTML table created by DataTables plugin.See Details.

fieldnames
The fieldnames(colnames) of the data table stored in .json file

tableParentNode
XMLInternalNode allow user to specify which node the table will be added to.

Details

When customFilter is set to "auto",it automatically create individual column filters based on data type of each col. For instance, if a column contains characters, it's column filter will be a search box; if a column contains numbers,it's column filter will be range filter; User can also specify which specific column(s) need to be filter. For instance, if user needs col4 of a five-column table to be filtered by keyword search, simply set the 4th column to "search", and the rest of cols to "null": customFilter = c("null","null","null","search","null") The output of this function is a .html file, along with required css and javascript library, as well as a data folder containing the .json file. User can copy the whole output folder to the web server to open the .html file.

Value

a .html file containing the customed javascript required for loading the table stored in .json file

[Package myHTMLtool version 0.2 Index]