addJsonFileToHTML {myHTMLtool}	R Documentation

Add .json file to HTML object.

Description

Add data table stored in .json format to the HTML object. Based on the data stored in .json file, a customed javascript will also be created and inserted into the <head/> node of HTML object.

Usage

addJsonFileToHTML(jsonFilePath, htmlObj, fieldnames, customFilter = NULL,
  tableParentNode = NULL)

Arguments

jsonFilePath
Path to the .json file

htmlObj
A html object returned by initHTML() #

fieldnames
The fieldnames(colnames) of the data table stored in .json file #

customFilter
A user-defined criteria to add individual column filtering in # the HTML table created by DataTables plugin.See Details. # #

Details

When customFilter is set to "auto",it automatically add column filters # based the data type of each col.if a column contains characters, it's column filter # will be "search"; if a column contains numbers,it's column filter will be "number-range"; # User can also specify which specific column(s) need to be filter. # For instance, if user needs col4 of a five-column table to be filtered by keyword search, # simply set the 4th column to "search", and the rest of cols to "null": # #

Value

a .html file containing the javascript required to load the table stored in .json file # #

[Package myHTMLtool version 0.2 Index]