addReportTemplate {myHTMLtool}	R Documentation

Add report template to htmlObj

Description

Add report headings to an empty htmlObj

Usage

addReportTemplate(htmlObj, headings = c("Introduction",
  "Method summary and results"), addLogo = TRUE)

Arguments

htmlObj
A html object returned by initHTML()

headings
vector a list of report headings to be added

addLogo
boolean If TRUE, a Gilead logo will be added to the report

Value

updated htmlObj

[Package myHTMLtool version 0.2 Index]