createVarColTypes {myHTMLtool}	R Documentation

create a java script array to store col types

Description

create a java script array to store col types

Usage

createVarColTypes(dataframe, customFilter)

Arguments

dataframe
A R data.frame

customFilter
Character or vector.If set to "auto", will automatically determin the filter type for each column based of data type.

[Package myHTMLtool version 0.2 Index]