myHTMLtool {myHTMLtool}	R Documentation

Generate a HTML file containing data table

Description

myHTMLtools create a HTML file containing customed java scripts, which accommondates the loading of large data set into a table.

[Package myHTMLtool version 0.2 Index]